+++ src/node/node.c
@@ -2,12 +2,12 @@
 
 struct node {
     unsigned short int target;
-    float weight;
+    double weight;
     struct node *next;
 };
 
 // Cria um novo nó
-Node *node_contruct(int target, float weight, Node *next) {
+Node *node_contruct(int target, double weight, Node *next) {
     Node *new_node = (Node *) malloc(sizeof(Node));
     new_node->target = target;
     new_node->weight = weight;
@@ -28,7 +28,7 @@ int node_get_target(Node *node) {
     return node->target;
 }
 
-float node_get_weight(Node *node) {
+double node_get_weight(Node *node) {
     return node->weight;
 }
